refactor(relay): group flat modules into parse/ sponsor/ effects/
The relay handler was a dozen flat files. Group them by concern so the structure
reflects the pipeline:

- parse/      { proposal, wallet } — decode the delegate action into a RelayOperation
- sponsor/    { mod=Sponsor, retry, policy } — the relayer: signing, retry, spend policy
- effects/    { accounting, registrations, background } — registrations + usage/metrics

Top level keeps the entry points and shared/cross-cutting pieces: submit (HTTP
handler), access (authorization), confidential (intents, has external callers),
sputnik (shared kind model), dto.

Pure move + import-path update; no behavior change. External paths
(relay::submit::*, relay::confidential::*) are unchanged.
